pyproject.toml
[project]
name = "insightvault"
version = "0.1.0"
description = "Multimodal manual-ingestion and RAG QA workflow with LangGraph"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "flagembedding>=1.3.5",
    "grandalf>=0.8",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.11",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.1.2",
    "loguru>=0.7.3",
    "magic-pdf>=1.3.12",
    "minio>=7.2.20",
    "modelscope>=1.35.0",
    "neo4j>=5.28.2",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pymilvus[model]>=2.6.10",
    "pymilvus-model>=0.3.2",
    "pymongo>=4.16.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "regex>=2026.2.28",
    "torch>=2.10.0",
    "uvicorn>=0.42.0",
    "transformers>=4.57.6",
    "openai-agents>=0.4.2",
    "requests>=2.32.5",
]
